bancoX-1.5v
The bank-X has a fresh new look compared to the previous version, along with new functionalities such as: account creation, user creation, and account listing.
Now, it's not possible to deposit negative numbers, and an error screen will be displayed every time an attempt to deposit negative or null numbers occurs. 
In certain sections, the symbol "===" indicates success, while the symbol "@@@" indicates an error.